+++ src/main/java/nextstep/roomescape/reservation/controller/dto/ReservationResponseDTO.java
@@ -1,10 +1,10 @@
-package nextstep.roomescape.reservation.domain.dto;
+package nextstep.roomescape.reservation.controller.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import nextstep.roomescape.reservation.domain.entity.Reservation;
-import nextstep.roomescape.reservation.domain.entity.Theme;
+import nextstep.roomescape.reservation.model.Reservation;
+import nextstep.roomescape.reservation.model.Theme;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
