+++ src/main/java/nextstep/roomescape/reservation/service/ReservationService.java
@@ -1,0 +1,12 @@
+package nextstep.roomescape.reservation.service;
+
+import nextstep.roomescape.reservation.controller.dto.ReservationResponseDTO;
+import nextstep.roomescape.reservation.controller.dto.ReservationRequestDTO;
+
+public interface ReservationService {
+    ReservationResponseDTO create(ReservationRequestDTO reservation);
+
+    ReservationResponseDTO findById(long id);
+
+    Boolean delete(long id);
+}
