+++ src/test/java/nextstep/roomescape/reservation/ReservationServiceImplTest.java
@@ -1,8 +1,10 @@
 package nextstep.roomescape.reservation;
 
-import nextstep.roomescape.reservation.domain.dto.ReservationResponseDTO;
-import nextstep.roomescape.reservation.domain.dto.ReservationRequestDTO;
-import nextstep.roomescape.reservation.domain.entity.Theme;
+import nextstep.roomescape.reservation.controller.dto.ReservationResponseDTO;
+import nextstep.roomescape.reservation.controller.dto.ReservationRequestDTO;
+import nextstep.roomescape.reservation.model.Theme;
+import nextstep.roomescape.reservation.repository.ReservationRepositoryMemoryImpl;
+import nextstep.roomescape.reservation.service.ReservationServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
